+++ NEWS.md
@@ -1,3 +1,15 @@
+# timekit 0.2.0.9000
+* Improvements:
+    * `tk_make_future_timeseries`: 
+        * Improved future date picking algorithm to look for seasonal trends by way of `inspect_months`. The data must span at least two calendar years to gain this benefit. 
+        * Improved handling of index in situations with units = days and scale greater than day (e.g. weekly, monthly, quarterly, +). Now values returned will be a regular series of the appropriate scale versus an irregular series indexed by regularly spaced seconds. For example, monthly dates in units = days will be predicted as the first day of each month rather than by the median frequency in seconds.
+        * Includes a new argument, `insert_values` to add time-based values into a time series. Before you could remove via `skip_values`, and now you can add via `insert_values`.
+        * Changed `n_future` to be inclusive of skip_values and weekends, which means that the end date of the future period is the same regardless of `inspect_weekdays`, `inspect_months`, `skip_values`, and now `insert_values`. Now the end dates are fixed which helps when manipulating future days.
+* Documentation:
+    * (New Vignette) TK02 - Making a Future Time Series Index using timekit
+    * (New Vignette) TK03 - Forecasting Using a Time Series Signature with timekit
+         
+
 # timekit 0.2.0
 
 * New Functions: